+++ financial_management/src/main/java/financial_management/parameter/wallet/WithdrawParam.java
@@ -8,14 +8,14 @@ package financial_management.parameter.wallet;
  **/
 public class WithdrawParam {
 
-    Long cost;
+    Double cost;
     String cardid;
 
-    public Long getCost() {
+    public Double getCost() {
         return cost;
     }
 
-    public void setCost(Long cost) {
+    public void setCost(Double cost) {
         this.cost = cost;
     }
